Moved construct_constraint_set in to Constraints
