+++ src/collisions/constraints.hpp
@@ -1,13 +1,19 @@
 #pragma once
 
-#include <vector>
-
+#include <ipc/collision_mesh.hpp>
 #include <ipc/collisions/collision_constraint.hpp>
 #include <ipc/collisions/vertex_vertex.hpp>
 #include <ipc/collisions/edge_vertex.hpp>
 #include <ipc/collisions/edge_edge.hpp>
 #include <ipc/collisions/face_vertex.hpp>
 #include <ipc/collisions/plane_vertex.hpp>
+#include <ipc/broad_phase/broad_phase.hpp>
+
+#include <Eigen/Core>
+
+#include <tbb/enumerable_thread_specific.h>
+
+#include <vector>
 
 namespace ipc {
 
@@ -18,6 +24,32 @@ struct Constraints {
     std::vector<FaceVertexConstraint> fv_constraints;
     std::vector<PlaneVertexConstraint> pv_constraints;
 
+    /// @brief Construct a set of constraints used to compute the barrier potential.
+    /// @param mesh The collision mesh.
+    /// @param V Vertices of the collision mesh.
+    /// @param dhat The activation distance of the barrier.
+    /// @param dmin Minimum distance.
+    /// @param method Broad-phase method to use.
+    void build(
+        const CollisionMesh& mesh,
+        const Eigen::MatrixXd& V,
+        const double dhat,
+        const double dmin = 0,
+        const BroadPhaseMethod method = BroadPhaseMethod::HASH_GRID);
+
+    /// @brief Construct a set of constraints used to compute the barrier potential.
+    /// @param candidates Distance candidates from which the constraint set is built.
+    /// @param mesh The collision mesh.
+    /// @param V Vertices of the collision mesh.
+    /// @param dhat The activation distance of the barrier.
+    /// @param  dmin  Minimum distance.
+    void build(
+        const Candidates& candidates,
+        const CollisionMesh& mesh,
+        const Eigen::MatrixXd& V,
+        const double dhat,
+        const double dmin = 0);
+
     size_t size() const;
 
     bool empty() const;
@@ -26,6 +58,39 @@ struct Constraints {
 
     CollisionConstraint& operator[](size_t idx);
     const CollisionConstraint& operator[](size_t idx) const;
+
+protected:
+    struct Builder;
+
+    static void edge_vertex_candiates_to_constraints(
+        const CollisionMesh& mesh,
+        const Eigen::MatrixXd& V,
+        const std::vector<EdgeVertexCandidate>& candidates,
+        const std::function<bool(double)>& is_active,
+        const size_t start_i,
+        const size_t end_i,
+        Builder& constraint_builder);
+
+    static void edge_edge_candiates_to_constraints(
+        const CollisionMesh& mesh,
+        const Eigen::MatrixXd& V,
+        const std::vector<EdgeEdgeCandidate>& candidates,
+        const std::function<bool(double)>& is_active,
+        const size_t start_i,
+        const size_t end_i,
+        Builder& constraint_builder);
+
+    static void face_vertex_candiates_to_constraints(
+        const CollisionMesh& mesh,
+        const Eigen::MatrixXd& V,
+        const std::vector<FaceVertexCandidate>& candidates,
+        const std::function<bool(double)>& is_active,
+        const size_t start_i,
+        const size_t end_i,
+        Builder& constraint_builder);
+
+    void merge_thread_local_constraints(
+        const tbb::enumerable_thread_specific<Builder>& local_storage);
 };
 
 } // namespace ipc
